Write model to file
When training is done we write the model to file
in a format expected by Ninja.
(see https://github.com/yuval/ninja)

This script learns a separate biases vector and weights
matrix for each layer. Ninja however expects both in the
same matrix, so we first combine them and then write a
single weights matrix for each layer.

For example: For a network with 784 input nodes,
2 hidden nodes, and 10 output nodes (784 2 10),
we first have the following matrices:
            w1 = 2 x 784
            b1 = 2 x 1
            w2 = 10 x 2
            b2 = 10 x 1

After combining biases and weights we have:
            w1 = 2 x 785
            w2 = 10 x 3

The model file is stored in model.txt